+++ lib/broker/fan_out.go
@@ -21,6 +21,7 @@ type FanOut struct {
 	logger log.Modular
 	stats  metrics.Type
 
+	maxInFlight  int
 	transactions <-chan types.Transaction
 
 	outputTsChans []chan types.Transaction
@@ -37,6 +38,7 @@ func NewFanOut(
 ) (*FanOut, error) {
 	ctx, done := context.WithCancel(context.Background())
 	o := &FanOut{
+		maxInFlight:  1,
 		stats:        stats,
 		logger:       logger,
 		transactions: nil,
@@ -54,6 +56,16 @@ func NewFanOut(
 		}
 	}
 	return o, nil
+}
+
+// WithMaxInFlight sets the maximum number of in-flight messages this broker
+// supports. This must be set before calling Consume.
+func (o *FanOut) WithMaxInFlight(i int) *FanOut {
+	if i < 1 {
+		i = 1
+	}
+	o.maxInFlight = i
+	return o
 }
 
 //------------------------------------------------------------------------------
@@ -159,7 +171,7 @@ func (o *FanOut) loop() {
 	}
 
 	// Max in flight
-	for i := 0; i < 50; i++ {
+	for i := 0; i < o.maxInFlight; i++ {
 		wg.Add(1)
 		go sendLoop()
 	}